test: place companion subprocess checks in the Cargo integration target
Preserve the opaque-wrapper usage and opt-out assertions with an isolated process environment. Declare existing shared-fixture dependencies for Cargo test compilation while retaining the locked developer tool versions.
