+++ crates/ctx-cli-qualification/Cargo.toml
@@ -13,8 +13,9 @@ publish = false
 autotests = false
 
 [package.metadata.cargo-shear]
-# Shared contract shims compile under Bazel and are not Cargo test targets.
-ignored = ["libc", "sha2", "uuid", "windows-sys"]
+# Shared contract modules live outside this package, which cargo-shear does not
+# follow. Their imports are required by the declared Cargo test targets.
+ignored = ["base64", "chrono", "ctx-history-core", "ctx-history-index", "flate2", "libc", "ring", "sha2", "tar", "uuid", "windows-sys"]
 
 [[test]]
 name = "index"
@@ -31,11 +32,18 @@ path = "tests/grok_build.rs"
 [dev-dependencies]
 anstream.workspace = true
 assert_cmd.workspace = true
+base64.workspace = true
+chrono.workspace = true
+ctx-history-core = { path = "../ctx-history-core" }
+ctx-history-index = { path = "../ctx-history-index" }
+flate2 = "1"
 libc.workspace = true
 predicates.workspace = true
+ring.workspace = true
 rusqlite.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
+tar.workspace = true
 tempfile.workspace = true
 uuid.workspace = true
