+++ crates/ctx-cli/src/companion/contract_tests.rs
@@ -849,69 +849,6 @@ fn blame_usage_root_preserves_non_utf8_native_paths() {
     );
 }
 
-#[cfg(unix)]
-#[test]
-fn paid_blame_wrapper_records_after_companion_exit_and_remains_controlled() {
-    use std::os::unix::fs::PermissionsExt as _;
-    use std::process::Command;
-
-    let temp = tempfile::tempdir().unwrap();
-    let pro = temp.path().join("ctx-pro");
-    std::fs::write(
-        &pro,
-        b"#!/bin/sh\nif [ \"$1\" = \"--ctx-pro-protocol-v3\" ] && [ \"$2\" = \"handshake\" ]; then\n  printf '{\"protocol_version\":3}\\n'\n  exit 0\nfi\n[ \"$1\" = \"--ctx-pro-protocol-v3\" ] && [ \"$2\" = \"cli\" ] && exit 0\nexit 91\n",
-    )
-    .unwrap();
-    std::fs::set_permissions(&pro, std::fs::Permissions::from_mode(0o700)).unwrap();
-
-    let enabled = temp.path().join("enabled");
-    std::fs::create_dir(&enabled).unwrap();
-    ctx_history_platform::platform_security::restrict_private_directory(&enabled).unwrap();
-
-    let status = Command::new(env!("CARGO_BIN_EXE_ctx"))
-        .arg("--data-root")
-        .arg(&enabled)
-        .arg("blame")
-        .arg("opaque-target")
-        .env("CTX_PRO_PATH", &pro)
-        .env_remove("CTX_LOCAL_USAGE_ENABLED")
-        .status()
-        .unwrap();
-    assert!(status.success());
-
-    let authority = ctx_client_observability::local_usage::LocalUsageStorageAuthority::new(
-        enabled.join("usage.sqlite"),
-        "1.0.0",
-    );
-    let report = ctx_client_observability::local_usage::read_report_authorized(
-        &authority,
-        &ctx_client_observability::local_usage::UsageControlSnapshot::unversioned(true),
-        true,
-    );
-    let definition = &report.definitions.unwrap()[0];
-    assert_eq!(definition.definition_version, 3);
-    assert_eq!(definition.summary.calls, 1);
-    assert_eq!(definition.summary.successful_calls, 1);
-    assert_eq!(definition.summary.not_applicable_calls, 1);
-    assert_eq!(definition.summary.result_count, 0);
-    assert_eq!(definition.summary.delivered_output_bytes, 0);
-
-    let disabled = temp.path().join("disabled");
-    std::fs::create_dir(&disabled).unwrap();
-    ctx_history_platform::platform_security::restrict_private_directory(&disabled).unwrap();
-    let status = Command::new(env!("CARGO_BIN_EXE_ctx"))
-        .arg("--data-root")
-        .arg(&disabled)
-        .arg("blame")
-        .arg("opaque-target")
-        .env("CTX_PRO_PATH", &pro)
-        .env("CTX_LOCAL_USAGE_ENABLED", "false")
-        .status()
-        .unwrap();
-    assert!(status.success());
-    assert!(!disabled.join("usage.sqlite").exists());
-}
-
 #[test]
 fn data_root_options_after_delimiter_remain_opaque_pro_arguments() {
     for trailing in [
